Fix a compiler warning.
Don't use the year given for a MythVideo file if it is set to the invalid
date of 1895. Also don't use the plot for the description if its is set to 
'None'.

git-svn-id: http://svn.mythtv.org/svn/trunk@9990 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mytharchive/mytharchivehelper/main.cpp
@@ -958,7 +958,7 @@ bool getFileInfo(QString inFile, QString outFile)
                 // calc duration of the file by counting the video frames
                 int duration = getFrameCount(inputFC, i);
                 VERBOSE(VB_JOBQUEUE, QString("frames = %1").arg(duration));
-                duration = duration / fps;
+                duration = (int)(duration / fps);
                 VERBOSE(VB_JOBQUEUE, QString("duration = %1").arg(duration));
                 root.setAttribute("duration", duration);
